export grouping functions (#10)
