+++ gpu_group.c
@@ -433,51 +433,6 @@ static int can_tile(__isl_keep isl_map *access, struct gpu_array_tile *tile)
 	return 1;
 }
 
-/* Internal data structure for gpu_group_references.
- *
- * scop represents the input scop.
- * kernel_depth is the schedule depth where the kernel launch will
- * be introduced, i.e., it is the depth of the band that is mapped
- * to blocks.
- * shared_depth is the schedule depth at which the copying to/from
- * shared memory is computed.  The copy operation may then
- * later be hoisted to a higher level.
- * thread_depth is the schedule depth where the thread mark is located,
- * i.e., it is the depth of the band that is mapped to threads and also
- * the schedule depth at which the copying to/from private memory
- * is computed.  The copy operation may then later be hoisted to
- * a higher level.
- * n_thread is the number of schedule dimensions in the band that
- * is mapped to threads.
- * privatization lives in the range of thread_sched (i.e., it is
- * of dimension thread_depth + n_thread) and encodes the mapping
- * to thread identifiers (as parameters).
- * host_sched contains the kernel_depth dimensions of the host schedule.
- * shared_sched contains the first shared_depth dimensions of the
- * kernel schedule.
- * copy_sched contains the first thread_depth dimensions of the
- * kernel schedule.
- * thread_sched contains the first (thread_depth + n_thread) dimensions
- * of the kernel schedule.
- * full_sched is a union_map representation of the entire kernel schedule.
- * The schedules are all formulated in terms of the original statement
- * instances, i.e., those that appear in the domains of the access
- * relations.
- */
-struct gpu_group_data {
-	struct ppcg_scop *scop;
-	int kernel_depth;
-	int shared_depth;
-	int thread_depth;
-	int n_thread;
-	isl_set *privatization;
-	isl_union_map *host_sched;
-	isl_union_map *shared_sched;
-	isl_union_map *copy_sched;
-	isl_union_map *thread_sched;
-	isl_union_map *full_sched;
-};
-
 /* Construct a map from domain_space to domain_space that increments
  * the dimension at position "pos" and leaves all other dimensions
  * constant.
@@ -883,7 +838,7 @@ struct gpu_array_ref_group *gpu_array_ref_group_free(
 /* Check if the access relations of group1 and group2 overlap within
  * copy_sched.
  */
-static int accesses_overlap(struct gpu_array_ref_group *group1,
+int accesses_overlap(struct gpu_array_ref_group *group1,
 	struct gpu_array_ref_group *group2)
 {
 	int disjoint;
@@ -898,7 +853,7 @@ static int accesses_overlap(struct gpu_array_ref_group *group1,
 /* Combine the given two groups into a single group, containing
  * the references of both groups.
  */
-static struct gpu_array_ref_group *join_groups(
+struct gpu_array_ref_group *join_groups(
 	struct gpu_array_ref_group *group1,
 	struct gpu_array_ref_group *group2)
 {
@@ -936,7 +891,7 @@ static struct gpu_array_ref_group *join_groups(
 /* Combine the given two groups into a single group and free
  * the original two groups.
  */
-static struct gpu_array_ref_group *join_groups_and_free(
+struct gpu_array_ref_group *join_groups_and_free(
 	struct gpu_array_ref_group *group1,
 	struct gpu_array_ref_group *group2)
 {
@@ -1197,7 +1152,7 @@ static int compute_group_bounds_core(struct ppcg_kernel *kernel,
  * reference group "group" of array "array" and set the tile depth.
  * Return 0 on success and -1 on error.
  */
-static int compute_group_bounds(struct ppcg_kernel *kernel,
+int compute_group_bounds(struct ppcg_kernel *kernel,
 	struct gpu_array_ref_group *group, struct gpu_group_data *data)
 {
 	if (!group)
@@ -1268,7 +1223,7 @@ static int group_overlapping_writes(struct ppcg_kernel *kernel,
 /* Check if the access relations of group1 and group2 overlap within
  * the outermost min(group1->min_depth, group2->min_depth) loops.
  */
-static int depth_accesses_overlap(struct gpu_array_ref_group *group1,
+int depth_accesses_overlap(struct gpu_array_ref_group *group1,
 	struct gpu_array_ref_group *group2)
 {
 	int depth;
@@ -1307,7 +1262,7 @@ static int group_depth_overlapping_writes(struct ppcg_kernel *kernel,
 /* Is the size of the tile specified by "tile" smaller than the sum of
  * the sizes of the tiles specified by "tile1" and "tile2"?
  */
-static int smaller_tile(struct gpu_array_tile *tile,
+int smaller_tile(struct gpu_array_tile *tile,
 	struct gpu_array_tile *tile1, struct gpu_array_tile *tile2)
 {
 	int smaller;
@@ -1407,7 +1362,7 @@ static void set_array_groups(struct gpu_local_array_info *array,
 /* Combine all groups in "groups" into a single group and return
  * the new number of groups (1 or 0 if there were no groups to start with).
  */
-static int join_all_groups(int n, struct gpu_array_ref_group **groups)
+int join_all_groups(int n, struct gpu_array_ref_group **groups)
 {
 	int i;
 
@@ -1494,7 +1449,7 @@ static int group_array_references(struct ppcg_kernel *kernel,
  * had their order dependences added to prog->array_order and
  * subsequently to the coincidence constraints.
  */
-static void check_can_be_private_live_ranges(struct ppcg_kernel *kernel,
+void check_can_be_private_live_ranges(struct ppcg_kernel *kernel,
 	__isl_keep isl_schedule_node *node)
 {
 	int i;
